Move useBreadcrumbTruncation to @tsmono/react/hooks
Both apps carried near-identical copies; the only real difference was
inspect wrapping the effect body in useCallback. Adopted scout's
cleaner shape (effect-local function plus an explicit guard for empty
first/last segments) and dropped the duplicates.
