+++ apps/inspect/src/app/navbar/Navbar.tsx
@@ -2,6 +2,7 @@ import clsx from "clsx";
 import { FC, Fragment, ReactNode, useMemo, useRef } from "react";
 import { Link } from "react-router-dom";
 
+import { useBreadcrumbTruncation } from "@tsmono/react/hooks";
 import { basename, dirname, ensureTrailingSlash } from "@tsmono/util";
 
 import { useStore } from "../../state/store";
@@ -9,7 +10,6 @@ import { prettyDirUri } from "../../utils/uri";
 import { ApplicationIcons } from "../appearance/icons";
 
 import styles from "./Navbar.module.css";
-import { useBreadcrumbTruncation } from "./useBreadcrumbTruncation";
 
 interface NavbarProps {
   children?: ReactNode;
