+++ apps/scout/src/app/components/BreadCrumbs.tsx
@@ -2,13 +2,13 @@ import clsx from "clsx";
 import { FC, Fragment, useMemo, useRef } from "react";
 import { Link } from "react-router-dom";
 
-import { basename, dirname, prettyDirUri } from "@tsmono/util";
-
-import styles from "./Breadcrumbs.module.css";
 import {
   BreadcrumbSegment,
   useBreadcrumbTruncation,
-} from "./useBreadcrumbTruncation";
+} from "@tsmono/react/hooks";
+import { basename, dirname, prettyDirUri } from "@tsmono/util";
+
+import styles from "./Breadcrumbs.module.css";
 
 const kPathSeparator = "/";
