+++ packages/react/src/hooks/useBreadcrumbTruncation.ts
